Move prompt builders to storyteller service
